refactor: extract language-neutral value adapter core
Move shared layout and planning contracts out of Rust adapter ownership.
Dispatch on the DWARF source language before recognition so C and C++
values never enter Rust standard-library adapter paths while preserving
the existing DWARF fallback.
