+++ ghostscope-dwarf/src/language/rust/mod.rs
@@ -3,12 +3,7 @@ mod plan;
 mod value;
 mod variant;
 
-pub(crate) use plan::{btree_value_read_plan, hash_table_value_read_plan, RustPlanContext};
-pub(crate) use value::{
-    CompositeStructFieldCapture, IndirectSequenceAddressing, IndirectSequenceKind,
-    ProjectedPathSegment, ProjectedStructPresentation, ProjectedValuePresentation,
-    ProjectedValueRequirement, RingSequenceLengthKind, ValueLayout, ValueLayoutResolution,
-};
+pub(super) use value::{ValueLayout, ValueLayoutResolution};
 
 pub(super) fn resolve_tuple_index(index: u32) -> crate::VariableAccessSegment {
     access::resolve_tuple_index(index)
@@ -27,4 +22,13 @@ pub(super) fn resolve_value_layout(
 
 pub(super) fn requires_dwarf_qualified_name(current: &crate::ResolvedType) -> bool {
     value::requires_dwarf_qualified_name(current)
+}
+
+pub(super) fn build_value_read_plan(
+    context: &dyn crate::language::adapter::ValueAdapterContext,
+    current: &crate::ResolvedType,
+    layout: ValueLayout,
+    type_module_path: Option<&std::path::Path>,
+) -> crate::Result<Option<crate::ValueReadPlan>> {
+    plan::build_value_read_plan(context, current, layout, type_module_path)
 }
